Code review: (1) made seeds work cross-platform; (2) adjusted role auth logic
