+++ package.json
@@ -12,7 +12,7 @@
     "db:migrate": "prisma migrate dev",
     "db:push": "prisma db push",
     "db:seed": "prisma db seed",
-    "db:seed:dev": "NODE_ENV=development prisma db seed",
+    "db:seed:dev": "cross-env NODE_ENV=development prisma db seed",
     "db:studio": "prisma studio",
     "db:export": "tsx scripts/export-problems.ts",
     "check:limits": "tsx scripts/check-field-limits.ts",
@@ -39,6 +39,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "cross-env": "^10.1.0",
     "eslint": "^9",
     "eslint-config-next": "^16.3.4",
     "tailwindcss": "^4",
