+++ prisma/seeds_development.ts
@@ -1,19 +1,5 @@
 import type { PrismaClient } from "@prisma/client";
-
-export const DEVELOPMENT_USERS = [
-  {
-    email: "test-admin@vibemathed.local",
-    name: "Local Test Admin",
-    pseudonym: "TestMongoose001",
-    role: "researcher",
-  },
-  {
-    email: "test-user@vibemathed.local",
-    name: "Local Test User",
-    pseudonym: "TestOtter001",
-    role: "enthusiast",
-  },
-] as const;
+import { DEVELOPMENT_USERS } from "../src/lib/dev-users";
 
 /// Development-only identities for exercising authenticated UI locally.
 ///
